More updates to make the code build without warnings on NetBSD/pmax 5.0.
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@5483 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/cpus/cpu_ppc_instr.cc
@@ -996,8 +996,9 @@ X(fctiwz)
 		else if (frb.f <= -2147483648.0)
 			res = 0x80000000;
 		else
-			res = frb.f;
+			res = (int32_t) frb.f;
 	}
+
 	*(uint64_t *)ic->arg[1] = (uint32_t)res;
 }
